fix: corregir UsuarioRepository según diagrama de BD

+++ src/main/java/com/julianjax/kinalapp/repository/UsuarioRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
-    //spring genera automáticamente el SELECT * FROM usuarios WHERE estado = ?
+    // Spring genera automáticamente el SELECT * FROM usuarios WHERE estado = ?
     List<Usuario> findByEstado(int estado);
 
 }
